スクリプトコピー系を1つのファイルに統合。 getshapeをGeneratePreprocessFile.pyに統合
